add GO task

+++ torchdrug_tasks/dataset.py
@@ -20,7 +20,7 @@ def load_data(task_name, mol_emd, protein_emd):
     data = np.load(data_file)
     for key in data.keys():
         data[key] = np.nan_to_num(data[key])
-        data[key] = data[key].astype(np.float32)
+        data[key] = data[key].astype(float)
     if task_name in ["DrugBank", "Davis", "KIBA"]:
         x1, x2, labels = [data[f"{x}"] for x in ["x1", "x2", "label"]]
         shuffle_index = np.random.permutation(len(x1))
